Move translation logic to rendering
When rendering views in standalone Layout Inspector we always
want the view to be centered in the screen.

To achieve this we were translating the coordinates of each view
node directly at node creation.

This change moves the translation only to the rendering part,
so that the coordinates in the model still reflect the screen
space coordinates seen on the device.

Test: DeviceViewContentPanelTest
Bug: NA
Change-Id: I5c6192847e6910c49c22812a274789f3c29302bb