Fix big zoom-to-fit on startup
When Compose Preview is opening for the first time,
both DesignSurface.resize() and Preview.onAfterRender
applies zoom-to-fit.

DesignSurface.resize() applies zoom-to-fit on a Preview that is not
rendered yet. This results on calculating a big scale number
when zoom-to-fit is called.

This change will stop calling zoom-to-fit from DesignSurface.resize()
whenever a Preview is opened for the first time.
Instead it will be Preview.onAfterRender to apply the zoom-to-fit
only once the rendering is completed.

Fixes: 368527727
Test: visually tested
Change-Id: I4d9e9e4c3dac830f34a68bcb0b2b760f19acd925