Refine system image loading
Use the version of loadSynchronously that calls the callbacks directly.
We were not getting the local images callback initially, because the
RepoManager loads its local images on construction, and then we add the
listener, and it will only fire on subsequent changes to the local
images. Thus, we had to wait until remote images arrived to get any
images. With this change we should almost immediately see local images.

Instead of sending a list, define a state object that can signify if
local and/or remote images are loaded, and if there's an error.

Bug: 369676165
Test: ISystemImagesTest
Change-Id: I085e172d082c6ed5674e8da975166e378a2d7540