Disable JUnit producers only if we should sync or attach models
If we are not a Gradle project, it's a bit rude to remove JUnit run
configuration producers from AndroidGradleProjectStartupActivity.
Only remove them if we are about to attach cached models or trigger
sync.

In tests, we create a project in the fixture before having it
populated with our test's files.  This means that the startup activity
does not trigger, so emulate the effect of it in the test setup.

Make an AndroidGradleProjectStartupActivityTest actually run the
activity before assessing its requirements.

Bug: none filed
Test: existing tests adjusted
Change-Id: Ibd99ee3b62998c78f3f0f0395db1a8d22b5aa15c
(cherry picked from commit 9957402258545417c6d25779b4c05d8985b5fe8c)