Extract user notification logic out of WHS Panel and into
... a WHSPanelController. The controller's lifecycle is tied to the
emulator controller's.

Before this change, we made the lifecycle of the WHSPanel tied to the
emulator controller's in order to display user notifications if the
panel was even closed once an action finished. When the theme is
changed, the panel for the previous theme is returned. However, when the
WHS popup is hidden, Intellij does not know that its UI should be
updated. This leads to the Panel's UI being stale.

With the current change, we keep the behaviour of notifying the user if
the panel is closed, but we also re-create the panel each time the popup
is shown. Thanks to this, the theme is always up-to-date.

Fixes: 369767252
Test: //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests
Change-Id: Ibd0c96c79899f3c9851ab75378fee5ef4ad02ab3