Introduce hard and soft limits in RenderExecutor
Before this change, the RenderExectuor queue had a hard limit of
having at most 50 tasks in its queue.

This could cause some cleaning actions to be dropped, potentially
causing memory leaks.

This change introduces two different limits to this queue, called
soft and hard, so that cleaning tasks are not considered in the
soft limit, but only in the hard limit.

Bug: 370941457
Bug: 370776377
Test: RenderExecutorTest
Change-Id: Iad0625a0ad937777bdc30b9337dffa5d9227a267