Fix KotlincWithQuickFixesParser
Apparently it was not accurate in replacing original build event
while adding quickfixes and description. e.g. event group information
was lost and it is used for analytics down the pipeline.

Also test was not very accurate, without real file present on disc
KotlincOutputParser produces slightly different message than normally.

This fixes and enhances the test with extra checks. Then enhances and
uses new events wrappers to add description and quickfixes.

In next change I will do the refactoring to merge this logic with
the one adding Gemini links.

Bug: N/A
Test: KotlincWithQuickFixesParser
Change-Id: Iddce52bd372aceacc61d99518b703a97738517bd