Use real dagger lib in tests
b/362992725 exposed a bug when running K2 that was not caught by tests.
That bug was fixed, but it was unclear at the time why the broken
scenario couldn't be captured in tests.

It turns out that there was a subtle difference in the definition of
the @Inject attribute that caused the K2 analysis engine to return
slightly different results. By using the real libraries in tests, we can
ensure we don't have discrepancies that creep in due to different
implementations of what should be the same class.

I've verified that the tests now would catch the same kind of error
previously seen in b/362992725 by undoing the fix and running with these
test updates; the tests would correctly fail in that circumstance.

Fixes: 364688659
Bug: 362992725
Test: All Dagger tests
Change-Id: I479d02c6fcf8ca5b0e2c0a91bcda9b051144ea72