Remove unnecessary code from model listener...
... in LayoutlibSceneManager.

The existing code in the modelChanged method is setting the need of
inflation to true, then requesting a render, and finally calling
selectionChanged after the render. But the last step is not needed
because when a inflation happens, the corresponding listeners are
called to make sure that the selectionChanged is properly invoked.

Bug: 361786751
Test: existing should pass
Change-Id: I4753d9e7abc3c929335b137a236a4b43b7f50133