Include Expected vs Actual Diff in Assertion Error
The `Generated` tests emit the `expected` and `actual` text to the test log but it's not included in the assertion itself. So when we view test failures in the logs or in Fusion, it's hard to find this information.

This cl replaces the assertion error that contains only the short message `Actual data differs from file content` with a comprehensive message that includes the entire  expected and actual texts as well as a diff generated by the Linux `diff` command.

Note:
I was not able to find a Java library that creates a human readable diff similar to what `diff` command does.

Example for what the output will look like:

https://fusion2.corp.google.com/invocations/7e343532-af53-47b1-945e-97a72fe0f0d6/targets/%2F%2Ftools%2Fadt%2Fidea%2Fij-debugger-tests:fails-art/tests

Bug: NA
Test: NA
Change-Id: I5e1a03eb3a155de3b0ab04dd908c1cfb5de755d9