Fix flaky "an asterisk only shows" test
The flakiness was caused because we were creating two separate
WHSPanels. One in the setup method and then another one with the
overridden applyChanges method. Both WHS panels were running coroutines
in parallel. When changing the textField values, this would trigger
a change in the capability state flow which would then update the text
field from the previously created panel's text field which would in turn
update the state flow, and so on. This led to a few rare instances where
the text field value would be overridden by a previous value. We'd end
up in an unexpected state which would result in the test timing out.

This CL changes the test class to only instantiate the WHSPanel once per
test case.

Bug: None
Test: //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests *10000
Change-Id: Ia8bc337223fc8480056580c9957d7ba9fbaaafdb