Limit coroutine scope lifecycle to the balloon's
The parent disposable was the emulator view which gets disposed whenever
the user changes emulator views or closes the emulator window.

The user could show multiple balloons before the emulator view was
disposed meaning we would end up with multiple panels running
coroutines in parallel which were updating the state manager
in parallel too.

With this change, we ensure that the coroutines get cancelled
whenever the balloon is disposed.

Bug: None
Test: //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests
Change-Id: Iff8b9333b3c5381b8a94209a487422fc16e0f4c9