Make Studio/Standalone rendering classes know
their siblings directly.

This is to enable Studio implementation use `BuildTargetReference`s
instead of `Module`s.

Rework rendering related interfaces and component instantiation in a way
that Studio/standalone-rendering components can talk to each other
without having to conform to a common interface where the interaction
is private to one or another implementation.

1. Move methods to get a class loader reference out of
   `ModuleRenderContext` to `RenderModelModule`;

2. Pass implementation references between implementations as
   implementation types;

3. Update usages (and update parameterized templates similarly)

4. Use `ModuleClassLoaderManager.Reference.close()` instead of several
   alternatives and make its implementation delegate the operation to
   the implementation of its instantiator.

Bug: 331381702
Test: n/a
Change-Id: Ifa78cddc67b220349e2caec939b5f4011aec29eb