Switch from Action to normal JButton
Action system updates the action only when the action can possibly be
shown to the user. This causes the context button to show up on insights
that were generated with the context only to disappear immediately after
it. This, however, ends up being a poor experience to see the button
"flicker" on the screen.
Switching to JButton allows us to control its visibility even when the
button is not actually visible to the user. Its visibility is controlled
by the state of the insight flow.

Fixes: 365994514
Test: updated
Change-Id: Ib230d5fdad69936c0ffb24a3f14ba4126658adaf