Make IDE recognize KMP modules
IDE would ignore kmp modules when calling moduleProjectTypeFromPlugin()
function. This fixes that.
Additionally, it removes the static calculation of the variant name and
instead uses the one coming from the build system.
Previously this would break for KMP which has no build type and flavour
but has a variant name (androidMain). However since the name was derived
from build types and flavours it ends up being an empty string and hence
throwing NullPointerExcetion in update() function.

Bug: 368481900
Bug: 371071431
Test: locally and updated unit tests
Change-Id: I9897658b656c31a53741f420aa2f8df0185e2ec5