Change PromptBuilder documentation to let attachments bypass aiexclude
Currently, the policy (which only exists in and is enforced by
documentation) that clients must declare all files they "used" in
prompts so that they can be checked against aiexclude is blocking
the attachment feature, because it needs to work with files outside
the project that the aiexclude service always treats as excluded.

Because such files are explicitly attached by the users, they shouldn't
be subject to aiexclude, which in spirit is meant to stop the IDE from
reading and sending the contents of files in the background without
the user's permission. So we're updating the documentation to reflect
this, so that the attachment feature doesn't appear to be violating
the policy.

Bug: 371212695
Test: n/a
Change-Id: Id115c613ff7e2ddc1c793c2e88eb82e7416061e1