Move compose preview AI actions out of compose-designer
This CL moves the following features related to the use of
Studio Bot into the Studio Bot plugin:
 - COMPOSE_PREVIEW_GENERATE_PREVIEW
 - COMPOSE_PREVIEW_GENERATE_ALL_PREVIEWS_FILE
 - COMPOSE_SEND_PREVIEW_TO_STUDIO_BOT

This is mostly so that the current ml-api module can be
removed entirely.

Some of these actions were fairly isolated, so it was a
straightforward move. But a custom extension point,
ComposeStudioBotActionFactory, has been added that allows
the Studio Bot plugin to contribute actions to the designer.

Bug: 327433971, 360363746
Test: existing
Change-Id: I1814adde5a777f8a8493c646baec83e7eeb19e44