Move usages of ResourceResolver to ResourceItemResolver
Moving a number of common cases from using ResourceResolver to
ResourceItemResolver.
ResourceResolver is expensive to create since it calculates all the
configured resources ahead of time. When we are only trying to lookup
one element, this is, in many cases not necessary.
ResourceItemResolver creates the ResourceResolver lazily when needed and
can do faster 1 element lookups.

Bug: 372888482
Bug: 322623569
Bug: 372679334
Bug: 370982774
Test: Existing
Change-Id: I27baa45e1a26ec4e4a1e066b0de415786790e9ac