[PSD] Avoid concurrent table modificiation on dispose
Using views on data structures during operations which might lead to
callbacks (such as disposers) being called is problematic: the
disposer might wish to modify the data structure being viewed, but
that's not allowed.

Generate an explicit list for the keys of `myConfigurables` for
operations that might trigger dispose.  (Strictly speaking this isn't
strong enough because one thread could in principle be trying to
iterate in a read-only manner over this structure while another tries
to dispose it, but in practice that's very unlikely while the PSD
remains modal.)

Fixes: 372902943
Test: caught by existing uitests
Change-Id: Ia2d49606cc3162be33b2538576408bcfe3b8c32e