Use the life-cycle's coroutine scope to initialize flows
... instead of the CommonPreviewRepresentation's

Each time the preview representation was activated, we initialized
the flow manager flows as well launched another coroutine. This was done
in the initializeFlows method. That method is called from the activate
method, which itself is called from the life-cycle's callbacks.

I mistakenly forgot to make the activate method an extension of
CoroutineScope in I8a25630a1e8fcb6a78701e9323. This meant that when
calling initializeFlows, the CoroutineScope that was being used was
the CommonPreviewRepresentation instead of the LifeCycle's. This was
an easy mistake to make as CommonPreviewRepresentation implements
AndroidCoroutinesAware which itself implements CoroutineScope. If it
weren't for that, the initializeFlows call from activate would have
a compilation error.

This caused a couple of issues:
1. a coroutine leak because each time the lifecycle was activated a new
flow was initialized which was never canceled in the de-activate.

2. because of the above, the preview flows of all activated
representations would re-refresh their previews whenever java/kotlin
code would change. This is because we have java/kotlin modification
trackers because a multi-preview could be declared in any java/kotlin.
Whenever we had multiple preview flows for the same representation, they
would end up analysing the same file concurrently, which might be
what was causing the freezes and locks.

By making the activate method be an extension method of CoroutineScope,
this makes the activate and then initializeFlows run under the
LifeCycle's coroutine scope. This scope is canceled on each deactivate.
This should ensure that the flows/coroutines don't leak, and ensures
that the preview providers don't run concurrently.

Bug: 370595516
Bug: 355667869
Bug: 344639845
Bug: 372643558
Bug: 372741826
Bug: 339651420
Test: existing + manual
Change-Id: I3c3530bcb53bf4e0dff1ec47878af8e4d9b211b1