[WORKAROUND] bypass CMakeUsageHandlerFactory
Currently, CMakeUsageHandlerFactory.canFindUsages will return true for
all PsiNamedElement. And it shows up before DefaultUsageHandlerFactory
and JavaUsageHandlerFactory. As a result, BUILD file and java file find
usage would use CMakeUsageHandlerFactory which cannot resolve many usage
correctly.

As we cannot remove this implementation directly, we create a find usage
factory to load at first and delegate to other factory with ignoring CMakeUsageHandlerFactory. KytheFindUsagesHandlerFactory also need to run the ignore check as it will be used whenever Kythe serivce is available.

We do not check the instance as import the class would make
KytheFindUsageHandlerFactory depends on intellij.cmake.psi which seems
to introduce some unnecessary deps to bypass this issue.

Bug: b/359391056, b/370573555
Test: manually
Change-Id: Iabbca7a98114fbe598d83c614b1c02be461a9ac0