Remove dubious quick fix from missing build tools checker
If gradle is failing because build tools are missing, it's probably
not a great idea to suggest an AGP upgrade at this point.  Remove the
hyperlink that attempts to do that AGP upgrade (which did reuse the
AGP Upgrade Assistant machinery, though through a distinct entry
point).

This also allows deletion of other dubious code, such as an attempt to
search for textual instantiations of AGP in build files, but only
through the "group:name" literal syntax and not any of the other
variants that have come up over the years.

Bug: 372751404
Test: existing
Change-Id: Ic4332915f03c08371c780781ea4e732e29427e7a