Rewrite GradleFiles updater in Kotlin
Allows use of coroutines in scheduling all the little bits of update
that need to happen.  Attempt to make all of that happen in apparent
parallel, while waiting for all the results to come in before reducing
and proceeding.

In more detail:

Define GradleFilesUpdater#computeFileHashes defines a number of helper
suspend functions to compute and store hashes and compute partial
results, for different areas of project build configuration: root
settings, module build files, version catalog files, gradle cache
files, gradle wrapper properties, etc.  All of these helpers are
launched under a Dispatchers.IO context, so that they will be
scheduled to run on io-aware threads.  This execution model is
intended to mimic the previous use of the BackgroundDiskIoExecutor.

We need to wait for the results of all of these jobs to be available
before computing the new GradleFiles state.  We can use #awaitAll(),
which has the benefit of also being a suspend function, so unlike in
the previous implementation, we do not impose an internal ordering on
these jobs: only that they should all be complete before reduction.
This is an improvement.  Also, all data structure manipulation is to
local data, so there is no need for synchronization between these
jobs.

Finally, after calculating the result, we can execute the callback
provided by GradleFiles, which makes the update to its
state (effectively) atomic.  GradleFilesUpdater#scheduleFileHashes is
the interface; we no longer need GradleFiles#updateFileHashes, and
GradleFilesUpdater#updateFileHashes is an internal method effectively
for allowing this to be executed effectively synchronously in tests.

Bug: 372751404
Test: rearranged slightly
Change-Id: I799816e2e62416f72b2c31dd1e9f8d5f46e469e4