Move registry storage into manager
GMavenIndexRepository is primarily responsible for managing the raw
index data, which is fetched periodically and cached on disk. It
complicates things to have it also manage the lifecycle of
MavenClassRegistry, which is effectively the parsed version of that
underlying data.

This moves the storage logic for MavenClassRegistry into
MavenClassRegistryManager. In addition to making intuitive sense, this
allows us to make more targeted fixes around how/when the registry is
created, since parsing it is a perf bottleneck.

This change is a functional no-op. The storage logic for the registry is
a little odd and has a few race conditions, but is identical to what was
there before. A following change will fix up that logic.

Bug: 373342815
Test: Covered by existing
Change-Id: I70b44a7349a72fe1080c4cd4e50044dcd848d3c8