Track deps with missing artifacts.
In some cases, it's possible for a build artifact from a dependency to
be missing (expired from the cache) when its needed. This never happens
in normal cases, since by the time the artigact expires from the cache
it will already be present in the artifact store. But there are cases
where it can happen, for exampe if some configuration is changed such
that the transform applied to an artifact is different.

Rather than failing the sync in this case, simply track which deps had
missing artifacts, and use that information when deciding whether or not
to enable analysis for a source file.

The upshot of this is that if artifact are missing, than analysis for
the dependant source file should be disabled again. The user can then
re-enable analysis for the file to get back to a good state.

This also helps ensure consisency in the event of config changes by
ensuring that artifact in the project store that correspond to the old
configuration are removed and replaced.

To do this, we have to make a small change to where the project artifact
store update is initiated from, since the result from that update can
affect the snapshot itself. Move the update call from ProjectUpdater to
QuerySyncProject to handle this.

Note that this CL fixes the crashes observed, but does not always result
in analysis being disabled for affected project sources. This is because
we use the query deps graph to track external dependencies whcih does
not include transitive external deps. So if we are missing artifacts for
a transitive (rather than direct) external dependency, we do now know to
disable analysis for the file. This can be fixed using information
already available to the IDE, and will be done in follow-up CLs.

Bug: 372069513
Test: com.google.idea.blaze.android.google3.qsync.SyncedPrebuiltProjectTest#missing_cache_artifacts

Change-Id: Id73452a6481e129ab50cc1d03c8f695199d25566