Improve table focus behavior
Previously, each table row was a focusable item, and the focus could
move indepedently of the selection. Pressing spacebar would then cause
the focused row to be selected. This is awkward, and forces you to tab
through every row in the table to move focus past the table.

Now, the table body is a single focusable element; rows are not
individually focusable. The up and down arrows change the table
selection whenever the focus is on the table or its header columns.

Bug: n/a
Test: DeviceTableTest
Change-Id: I5710c8c660bd639ba79762e65b83c0b0aaccbbf7