Use 'try' getter for import action
Replacing a lot of blocking calls to get the Maven registry with
non-blocking "try" calls.

The new API will only return null the first few times it's called. The
first call triggers the registry to load in the background, and once
that's complete the first time there will always be a value returned.

The locations being updated in this change all related to quick fixes
shown in the editor. If a quick fix isn't immediately available, it will
be shown shortly thereafter, since the editor queries for quick fixes
early and often.

Fixes: 373342815
Test: Included
Change-Id: Ib3ca5c1cd3c9569be2a37a6f8df1e3b15dd01b81