Update Gradle completion to handle to not block for gmaven index
Completion has to return results quickly. If the GMaven index hasn't
been loaded yet, this logic will no longer wait for it to be loaded and
instead will not provide values. Loading will be triggered in the
background at that point, and once it completes the index will be
available for the lifetime of the Studio process.

Bug: 373342815
Test: GradleDependencyCompletionContributorTest
Change-Id: Id79d61cced92f1ebf5c456a699e01e31210f9bce