Remove usages of com.intellij.dom.fileDescription
This EP is deprecated. It can be directly replaced with dom.fileMetaData
with no changes. However, for descriptions that only handle a single root
tag, specifying that in the registration allows for significant
performance improvements.

The descriptions that do not have a root tag name specified are those
that can handle multiple tags, so the omissions are deliberate.

Fixes: 374080683
Test: Covered by existing DOM highlighting and autosuggest tests
Change-Id: I1a9adef2cac9091d1b58f56671aa47252eb77ab6