Partially un-fork ErrorTooltip
It turns out that we don't need to fork Compose's TooltipArea; if we
don't emit any nodes in the tooltip composable, it doesn't display
anything. We just need to fork Jewel's Tooltip to avoid emitting the
empty box when there's nothing inside it.

This also avoids a bug where the tooltip lingers inappropriately.

Fixes: 374214668
Test: n/a
Change-Id: I7881cdf9624107e2b53b8d93f185dcb30f7e8ad0