Update studio gemini onboarding screen
to use the design we made for firebase Tos, according to latest UX mocks

Also refactored and improved the onboarding feedback system to make it:
1) generic so that it can be implemented by each product
2) support enabling or disabling the action button

In the future, we can expand the onboarding provider to support
composite actions. For example if we were to show the firebase Tos, we
would require users to onboard both Gemini and Firebase, in that order.

Test: new test.
Fixes: 374100331
Fixes: 373303476
Change-Id: I8d3e09ae527a2cacdbb0acfacd94bbe4b8e58d0c