Replace deprecated usages in AndroidOpenFileAction
Unfortunately the new APIs require Kotlin, so I have to create a utility
file in Kotlin to call them from Java.

The old APIs are being removed because they're deprecated. But the new
`doOpenProject` and `OpenProjectTask` APIs are marked as @Internal. We
have to use them to maintain current functionality, so I suppose that's
the lesser of two evils (as compared to using a deprecated API doing the
same thing that will be removed).

Bug: n/a
Test: n/a
Change-Id: Ic1993d27a8787faa8098b4a52737bda80dda866e