Move DynamicAppUtils
In theory these are no-longer Gradle-specific, so move them to a more
generic utility directory.

In practice some of the logic is not generally portable to all build
system project representations, because (for example) it operates at
the level of modules and has as an implicit assumption that a module
can be assigned a single "kind" of Android delivery entity (e.g. a
dynamic feature), as opposed to multiple entities within the same IDE
Module.  This implicit assumption is currently approximately valid for
Gradle-based projects, and not for projects based on other build
systems.

Support for other build systems can be implemented as and when desired
by adding to the methods in DynamicAppFeatureOnFeatureToken.

Bug: 300434824
Test: existing
Change-Id: Ib22cb947214b6b13d2a526a62793449c7333c345