Move GradleLocalCache
In principle this is not actually Gradle-specific at all: it
interrogates the developer's local Gradle cache for artifacts and
other information, but there's no reason that that necessarily happens
for Gradle-based projects only.

In practice, it's only used by the Gradle project system, so move it
out of temptation's way.

Bug: 300434824
Test: existing
Change-Id: Ic56cf03efebbc37e93214995cdcf8f9a268e25c9