Make BuildTargetReference not hold module reference
but only a weak reference. This is make it a replacement for weak
module references used in rendering.

Weak references are used to prevent leak detectors in tests from
complaining about references to modules held from the layout lib.

In Bazel implementation there no tests for now and thus a direct module
reference is fine as long as the implementation correctly returns `null`
when the module is disposed.

Bug: 331381702
Test: n/a
Change-Id: Ie1a9c22d78a5c3c2920db78521d72b602c0674ae