llm prompt: retain files used during construction of the prompt
The `buildLlmPrompt` API did not retain the set of files used
during the construction of the prompt. As a result, during the
conversion from the LlmPrompt to Prompt data type, we use
an empty list of files as context. Even though this works for now,
it is safer to just retain the files in case the implementation
of Prompt changes later.

Bug: 360363746
Test: updated
Change-Id: I8cee72a19838753e12b07bb8e4b349aa21207ef1