[AUA] Refactoring to block upgrade if aidl property is applied
AIDL property is being removed from Gradle properties in AGP 9.0. We
need to block upgrades to 9.0 on projects that apply the property (it is
present and not set to no op).

This changes introduces AbstractBlockPropertyUnlessNoOpProcessor that
blocks upgrades that crosses a given version when a property is present
and performs an operation.

This abstract class is extended for android.defaults.buildfeatures.aidl
and AGP 9.0.0-alpha01 and blockins upgrades from 8.0.0-alpha4 or older
to prevent conflicts with AidlDefaultRefactoringProcessor.

Bug: 374156390
Test: BlockAidlPropertyPresentRefactoringProcessorTest
Change-Id: Icadf572798856e1a295ebe024f6530c44edd0128