Adjust size of table columns
Introduce a TableColumnWidth that uses TextMeasurer to determine the
with required to display a given bit of text. Use it for all but the
Name column.

Making most column widths unweighted revealed a discrepancy between
header and row cell width calculation (previously hidden by being
divided among all of the weighted cells).

Bug: 375440141
Test: existing
Change-Id: I5711c5f2f36d45943e818561d8caeea31c143834