Remove stored zoom when zoom-to-fit
Whenever we switch from a layout option to another we store a scale.
When zoom-to-fit we apply we have different values for each layout
option.

Switching to a layout to another will refresh the previous store scale
value. Because zoom-to-fit is not a specific value, the scale applied in
the previous layout option could not be the zoom-to-fit value of the
next one.

Whenever we apply zoom-to-fit we delete the store value, in this way
whenever we switch layout option we calculate a new zoom level by
calling restoreZoomOrZoomToFit().

This fixes zoom-to-fit unconcistencies when changing from Grid
to Grid (Experimental).

Bug: 373348063
Test: existing tests should work
Change-Id: I18c68ce649b6119313c4310bb2a1a8b2ad2d2b13