Fix RoomSQL failure in advance of 2024.3 merge
The usage of `sourceElement` in question is actually marked with a
platform inspection, saying that it shouldn't be used:

> Usage of UElement as PsiElement is not recommended
> The PsiElement obtained this way is ambiguous.

It goes on to recommend two different ways we might fetch it. I'm
replacing the call with the appropriate variant.

This is working as expected right now, but will break with the merge
because the "ambiguous" API is returning something different. The fix
works both before and after the merge.

Fixes: 376053437
Test: TableReferencesTest.test_duplicatedTableNames
Change-Id: I54796e3aebc87b4ee86c14e9e0d53cf97bd6de1b