Fix DesignerEditorPanel global shortcut handling
Setting the PlatformCoreDataKeys.EDITOR in the split editor would
automatically redirect all global shortcuts to the editor side.
This change ensures that this does not happen. We also need to set the
IdeView since, if we do not have EDITOR, this is the way to inform the
IDE about the relevant options in the File menu.

Fixes: 376131411
Fixes: 376028920
Test: NavGraphSanityTest
Change-Id: Id3ad52455dbfa80d5a0f45491538d0291fa8b0e8