Fix infinite zoom scale when changing layout
When switching device types or when enabling modes (such as colorblind
mode) zoom-to-fit scale value was infinitive.

The reason was because a race condition between update sizes of
DesignSurface and the rendering of the content.

Applying waitForRenderBeforeRestoringZoom(true) will help handling the
race condition issues keeping also consistent the zoom level of the previous preview.

Fixes: 375584768
Fixes: 374663154
Test: visually tested
Change-Id: Icedcb2b5fbbe8c6d03415ee61c241901e5cd6d84