Unblock compose plugin K2 testing for 2024.3 merge
This test has a failure causing the semaphore never to be released.
Unfortunately, that's the second test to be run in the entire suite, and
the blockage causes the remainder of the tests to never even run.

This change doesn't fix the issue in the test with the 2024.3 merge (or
issues in some other tests for that matter), but it at least allows them
to start running so we can see and address the actual failures.

Bug: 376053698
Test: ComposableFunctionExtractableAnalyserTest.testComposableFunction
Change-Id: I04c23395d640cd8cc6925c47fe98a4781ca7808f