Launch Preview refresh process in the worker thread
Before this CL, the preview refresh process was being launched in
the UI thread, and the heavy parts of this processed were executed
under a context switch in the worker thread.

However, a very few sections of the whole refresh process actually
need to be executed in the UI thread, so this CL changes the approach
to launch in the worker thread and switch to the UI thread only
for those parts.

Also, some improvements related to context switching are being done.

Bug: 335424569
Test: existing should pass and manually verified Preview tools.
Change-Id: Ibe2527cf709e728aa9bed41661ed02e68da00abd