Adjust @Composable function completions for 2024.3
With K2 in 2024.3, some @Composable functions have two variants in the
autosuggest dropdown that end up looking identical after we rewrite
them. This change excludes one of them in that case.

There are also a few minor fixes to lookup insertion, to ensure that it
remains the same post-merge in both K1 and K2.

Bug: 376053698
Test: ComposeCompletionContributorTest
Change-Id: I787e861118c6bbbde190b3a337222ef4b7007fc6