Split EmbeddedDistributionPaths
All the gradle-related and gradle-adjacent methods are essentially
static, even if some are not currently expressed as such.  Make them
static, and move them to GradleProjectSystemUtils, leaving notionally
build-system-agnostic methods (related to the profiler bytecode
transformer and to Sdk locations) in the service class.

The remainder of the methods are also essentially static, but preserve
the service nature for now to minimize code churn.

Bug: 300434824
Test: existing tests adjusted
Change-Id: I1f7cc01a01c57486e34167ba06f050cd9c0c996d