Stop expecting single failure for KTS invalid build file
With IntelliJ 2024.3 the classpath mode was enabled by default for
import on KTIJ-31334 since previously was disabled because Gradle
swallowed important stacktraces during the import, making the
build jvm argument '-Dorg.gradle.kotlin.dsl.provider.mode=classpath'
to be specified from KotlinDslScriptModelResolver.

However, causes given a project with an invalid KTS build file to no
longer thrown from PhasedActionAwareConsumerConnection the
BuildActionFailureException and instead ProcessCanceledException is
thrown, resulting the process to be canceled and as a consequence the
build output messages are different.

This will be reported separately, but until then, the test was adapted
to no longer expect a single message.

Bug: 375693518
Test: KtsBuildFileCompilationBrokenTest
Change-Id: I515518324c2f2df6959aa342b20e64a6243b3968