Re-enable testIntentionPreviewConvertToDp
With upstream commit 40585d1a2592, the execution of a ModChooseAction
command is split to include a non-blocking read action for computing
the choices, before finishing the rest of the invocation on the UI
thread. In tests this means we need to add an extra wait, otherwise
the test might finish before the action is even called.

We only have one Lint quickfix that asks the user to choose between
mutiple options with a ModChooseAction, that is ConvertToDpQuickFix,
so it explains why only one particular test became flaky.

Bug: 376924928
Test: PreviewFixTest
Change-Id: Iaae7c175cd02b4a3fa7f18f2fc383bea0e3ccb03