[Dsl] Handle Project properties a little better
In particular, create Dsl entities in a way that matches the way that
the parser generates them from file: so, adding a new entry to an
existing ProjectPropertiesDslElement, rather than a top-level
GradleDslMethodCall with a strange name.

This means that we have to special-case in maybeTrimForParent, which
would otherwise remove the pseudo-redundant `project(':foo').` part of
the method call name, and in needToCreateParent, which should never be
creating elements for the ProjectPropertiesDslElement "block".  The
already-existing special-case logic over whether a
ProjectPropertiesDslElement should be ignored for anchoring purposes
also requires refinement.

This allows uncommenting some long-commented test code.

Bug: 158066552
Test: existing tests strengthened
Change-Id: I26ee8d44fa53cd4aa9e12aa4e8bb84afd6deaf93