[Dsl] Fewer newlines in Groovy files
The newline addition was quite eccentric, with some heroism (not
always successful) to attempt to avoid adding blank lines in blocks.
This change makes the Groovy formatting match the KotlinScript
formatting a little more, attempting to avoid adding newlines except
where necessary.

The test for when adding elements to a GrClosableBlock reflects the
fact that the Psi structure for
```
{
}
```
is approximately
```
GrClosableBlock
  {
  \n
  GrParameterList("")
  }
\n
```
Such that the parameter list is after the whitespace; statements are
added after that.

It's perhaps worth knowing that GrClosableBlock (and indeed
GroovyFile) implement GrStatementOwner and therefore have an
addStatementBefore() method, which might involve less fighting with
whitespace than we currently do... but it is *Before* and therefore
we'd need to rework how we manage our anchors.

Bug: 158066552
Test: existing tests eyeballed and adjusted
Change-Id: I80bc95d5bd66eecb3a85684cdb2520ae03b83e66