Use the specific tutorial bundle to look up images
Previously, ag/16332964 added support for looking up resources in a
assistant plugin's classloader, but didn't extend that to images.

This resulted in b/377421788 where images in the Android SDK Upgrade
Assistant were not visible in release builds.

This commit uses `myResourceClass` to fetch the image like the other
resources, which utilizes the correct class loader in release builds.

Fixes: 377421788
Test: verfied manually by checking behavior in a release build
Change-Id: I1416e7a2ad1cb5f2cd2b4c1f0a966afedf5d3020