[Dsl] Enrich our representation of anchors
The requestAnchor() method encodes an anchor to be used with
addAfter() as a DslElement, with `null` representing the null anchor,
with the semantics of "add at the start".  That's fine, and is capable
of representing everything we need.

However, getAnchor() also uses `null` as an error case (the case where
the element has no parent).  This leads to confusion -- at least mine,
and probably others'.

Encapsulate the return values of requestAnchor in a new (sealed)
interface, with a data object to represent the Start, and a data class
to capture the After case.  This allows `null` to be treated
differently from "add at start".

This change also clarifies a few places where we have dead code.

Bug: 158066552
Test: existing (no change in functionality, yet)
Change-Id: Ia8282f40078cbf16f54e7ad55d62cc456545e04c