Refine heuristic for V1 modules
The JSON plugin used to be a V1 platform module, thus its plugin ID
has the "com.intellij.modules" prefix usually reserved for V1 modules,
and this fools our logic in ValidatePluginConfigurationAction.

To handle this corner case, we can check the plugin ID in the
dependency plugin descriptor. (If the dependency name differs from
its containing plugin descriptor, then we know it must be a module
within some containing plugin.)

Bug: 376891148
Test: //tools/adt/idea/studio:StartUpTest
Change-Id: I477b528e4b6ed412564a136dd6726d2f2795689e