[Dsl] Place an added block after any existing similar elements
Rather than before.  This is in principle safer, because we could have
added elements which refer to existing ones.

This requires modifying the handling of Ext blocks in Kotlin, where
when we create it we need to return something that says that this
element exists without giving that element a PsiElement, which would
otherwise imply wrong things about the existence of a physical block.

(Fundamentally this is a multiple dispatch problem: we have at least
dispatch dimensions along the object identity of the element itself,
its parent, and the external language, with non-trivial pairs of these
being relevant for various operations and no obvious hierarchy between
them.  Short of writing an entire object system and re-implementing
everything in there, ad-hoc multiple dispatch like in the new
overriding methods on ExtDslElement is the best I can come up with.)

Bug: 158066552
Test: existing tests adjusted
Change-Id: I6e9c3373a6f67fe0906cee6d633559e454e823f6