Don't fetch full status in SafeArgs REx until needed.
The `getShadowedScope` method is called during LLFirSession creation,
including, among others, sessions created on the EDT to handle
modifications made when a write action ends. Previously, we checked
whether `currentStatus` was `null` to disable scope shadowing for
disabled REx instances, which had the unfortunate effect of completely
fetching the current status, causing EDT latency.

To fix this, limit the check to only looking at `NavInfoFetcher` status
(this matches the check made in `SafeArgsResolveExtensionProvider`.)
This check does not need to completely load and cache nav status, and
should be much quicker to execute from the EDT.

Fixes: 377859339
Test: bazel test //tools/adt/idea/nav/safeargs/tests/k2:all
Change-Id: Ie0091022a01a2f73517ea19f9c2de269cc55b12c