Add caching to our documentation target provider
This change reuses a read-through file based cache of URL fetches
I wrote for the device mirroring benchmarking. Ideally we would move
this to somewhere more general, but I haven't figured out where that
is yet. For now I will just fork and update it with a plan to
deduplicate later.

Bug: N/A
Test: AndroidSdkDocumentationTargetProviderTest.kt, UrlFileCacheTest.kt
Change-Id: I20286cc64d7f081212beaf258413ac59a6db38b6