Permission issue error dialog
Importing project requires write permissions. If folder has no such,
show a warning to user so they can first take care of that and try
again later.

If user decides to open despite this, we are going to extend catch
clause to show not only IOException but any Exception. Thanks to this,
if the app crashes due to lack of permissions - there will be error
pointing at which file or directory was the one involved.

ExternalSystemException appends lower level error message to the message
we use when creating it. For java.nio.file.AccessDeniedException that is
the file path. For that reason we are better passing only the file name,
to not impair readability by showing the path twice.

Bug: 369069561
Test: AndroidGradleProjectOpenProcessorTest.kt
Change-Id: I67f9ee1e0b56298719988b447651a17bbbe9622c