Fix deadlock in test
Given an unlucky sequence of actions, it is possible for a running
readAction to block an incoming write action, but that write action
also to block a second read action required to complete the first read
action.  This is a somewhat artificial case in this instance,
suggesting an issue in the way the test was formulated, but it is
related to other deadlocks in the system.

Fix this test by ensuring that the running read action is taken with
write intent on the EDT, which should prevent anything which actually
wants the write lock from being scheduled until after the test is
complete.  (This now survives 400 runs from the IDE).

Fixes: 379261305
Test: existing
Change-Id: Ia8c938757e34e976ed3fa44d8538e6ace56d5e04