Remove last use of Gradle Project System from android designer
Replace GradleBuildState.isBuildInProgress with
ProjectSystemBuildManager.isBuilding.  The latter is formally
deprecated, but I think its deprecation is motivated by issues related
to changes between the time-of-check and the time-of-use, and that is
no better in the GradleBuildState implementation.

This allows removal of the Gradle Project System dependency from the
android.designer module (and the consequent removal of that module
from the Gradle Project System allowlist).

Bug: 147586218
Test: existing
Change-Id: I76c625e5363edc20f560c6919d6b542fd8841494