Move OutOfProcessCompilerDaemonClientImpl
It is only used by test code at present.  Move it to the test,
reducing the amount of Gradle-specific code in compose-designer.

(The quantity of really Gradle-specific code is not huge: use of one
utility in GradleClassFinderUtil to intuit classpath and friend
directories to give as parameters to the out-of-process compiler; if
needed it could probably be reimplemented using tokens, possibly via
the BuildTargetReference argument.  But since it's unused in practice,
moving it away is simpler.)

Bug: 328608974
Test: existing
Change-Id: I30479e334b95b2c13522af3230be5a555e4931a1