Fix listener-removal bugs identified by the SAM detector
This CL fixes a handful of cases where listener lambdas
were passed to a functional interface so Kotlin applies SAM
conversion, creating a new instance at the usage site.
This doesn't work for listeners; this means that when you
attempt to remove the listener, you're passing in a different
instance than the one you registered earlier. The fix is
to change the listener variable to already use the functional
interface such that there's no SAM conversion.

(These cases were identified by the lint SAM detector, which
was just expanded to cover more scenarios.)

Bug: Not filed
Test: Existing
Change-Id: I2d6a18a85fc77173d67ba7e4aecb34a97af535d0