Rename kotlin.Metadata in JvmAnnotationNames
As part of the class loading, we re-package all the kotlin classes.
Unfortunately, kotlin.Metadata is reference by name in some of the
kotlin internals so moving it causes reflection to fail to load the
information stored in it.
This renames the constants that reference to kotlin.Metadata to ensure
that the class is loaded correctly.

Fixes: 240601093
Fixes: 281014894
Test: StringReplaceTransformTest
Change-Id: I29dd88bd7db1f8c6f57720938d093eb6d038b51c