More careful locking in ProjectSystemService
We are caught between a rock and a hard place: arbitrary IDE
subsystems can request the project system for that project, and
project system detectors might require a read lock -- but that read
lock can cause a deadlock through cycles in lock order.

Sail between the Scylla and Charybdis of threading problems by
admitting the possibility of a temporary wrong answer: if we are
attempting to compute and cache the project system at a point where
the read lock is unavailable, then use a default project system for
now; as and when we later compute the final project system, send roots
changed events to allow anything which is sensitive to the project
structure changing to recompute whatever is needed.

Bug: 375355918
Test: ongoing
Change-Id: I4669fcb5ee8bc9565eb2a833a1871ea1c1127232