Show split controls in a floating toolbar when editor tabs
... are hidden. After migrating to the new UI, the split editor controls
were moved to the editor tabs. Some users however, hide the editor tabs
to save vertical space. As a result, they can't see the split controls
and need to rely on remembering the corresponding shortcuts.

This is being changed so we display a floating toolbar whenever the user
hides the editor toolbars. This way, vertical space is preserved and the
actions can still be triggered using the mouse.

Fixes: 326091148
Test: Visually verified.
Change-Id: I1ac3e7cba7a72ea2d2d0ab991b94c15e82dbb476